feat: 게시글 api연동 및 디테일페이지 연결  로직 구현
